+++ debugger/src/dap_types.rs
@@ -54,8 +54,8 @@ pub struct ShowMessageEventBody {
 
 impl TryFrom<&serde_json::Value> for ReadMemoryArguments {
     fn try_from(arguments: &serde_json::Value) -> Result<Self, Self::Error> {
-        let count = get_int_argument(arguments, "count", 1)?;
-        let memory_reference = get_string_argument(arguments, "memory_reference", 0)?;
+        let count = get_int_argument(Some(arguments), "count", 1)?;
+        let memory_reference = get_string_argument(Some(arguments), "memory_reference", 0)?;
         Ok(ReadMemoryArguments {
             count,
             memory_reference,
@@ -69,8 +69,11 @@ impl TryFrom<&serde_json::Value> for ReadMemoryArguments {
 // SECTION: For various helper functions
 
 /// Parse the argument at the given index.
+///
+/// Note: The function accepts an `Option` for `arguments` because this makes
+/// the usage easier if no arguments are present.
 pub fn get_int_argument<T: Num>(
-    arguments: &serde_json::Value,
+    arguments: Option<&serde_json::Value>,
     argument_name: &str,
     index: usize,
 ) -> Result<T, DebuggerError>
@@ -81,7 +84,7 @@ where
     <T as Num>::FromStrRadixErr: 'static,
 {
     match arguments {
-        serde_json::Value::Array(arguments) => {
+        Some(serde_json::Value::Array(arguments)) => {
             if arguments.len() <= index {
                 return Err(DebuggerError::MissingArgument {
                     argument_name: argument_name.to_string(),
@@ -102,12 +105,12 @@ where
 }
 
 fn get_string_argument(
-    arguments: &serde_json::Value,
+    arguments: Option<&serde_json::Value>,
     argument_name: &str,
     index: usize,
 ) -> Result<String, DebuggerError> {
     match arguments {
-        serde_json::Value::Array(arguments) => {
+        Some(serde_json::Value::Array(arguments)) => {
             if arguments.len() <= index {
                 return Err(DebuggerError::MissingArgument {
                     argument_name: argument_name.to_string(),